MODFISTO-109 Add ability update encumbrances based on order_transaction_summary
- fix code smells
- remove duplicates
- refactoring
