+++ src/main/java/org/folio/dao/transactions/BaseTransactionDAO.java
@@ -1,13 +1,16 @@
 package org.folio.dao.transactions;
 
 import static org.folio.dao.transactions.EncumbranceDAO.TRANSACTIONS_TABLE;
-import static org.folio.rest.util.ResponseUtils.handleVoidAsyncResult;
+import static org.folio.dao.transactions.TemporaryInvoiceTransactionDAO.TEMPORARY_INVOICE_TRANSACTIONS;
+import static org.folio.rest.persist.HelperUtils.getFullTableName;
 import static org.folio.rest.util.ResponseUtils.handleFailure;
+import static org.folio.rest.util.ResponseUtils.handleVoidAsyncResult;
 
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
+import org.apache.commons.lang3.StringUtils;
 import org.folio.rest.jaxrs.model.Transaction;
 import org.folio.rest.persist.DBClient;
 import org.folio.rest.persist.Criteria.Criterion;
@@ -21,6 +24,9 @@ import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 
 public abstract class BaseTransactionDAO implements TransactionDAO {
+
+  public static final String INSERT_PERMANENT_TRANSACTIONS_BY_IDS = "INSERT INTO %s (id, jsonb) (SELECT id, jsonb FROM %s WHERE id in (%s)) "
+    + "ON CONFLICT DO NOTHING;";
 
   @Override
   public Future<List<Transaction>> getTransactions(Criterion criterion, DBClient client) {
@@ -80,7 +86,12 @@ public abstract class BaseTransactionDAO implements TransactionDAO {
 
   protected abstract String createPermanentTransactionsQuery(String tenantId);
 
-  protected abstract String createPermanentTransactionsQuery(String tenantId, List<String> ids);
+  protected String createPermanentTransactionsQuery(String tenantId, List<String> ids) {
+    String idsAsString = ids.stream()
+      .map(id -> StringUtils.wrap(id, "'"))
+      .collect(Collectors.joining(","));
+    return String.format(INSERT_PERMANENT_TRANSACTIONS_BY_IDS, getFullTableName(tenantId, TRANSACTIONS_TABLE), getFullTableName(tenantId, TEMPORARY_INVOICE_TRANSACTIONS), idsAsString);
+  }
 
   @Override
   public Future<Void> updatePermanentTransactions(List<Transaction> transactions, DBClient client) {
