+++ src/main/java/org/folio/dao/transactions/PendingPaymentDAO.java
@@ -22,11 +22,6 @@ public class PendingPaymentDAO extends BaseTransactionDAO {
   }
 
   @Override
-  protected String createPermanentTransactionsQuery(String tenantId, List<String> ids) {
-    return null;
-  }
-
-  @Override
   protected String buildUpdatePermanentTransactionQuery(List<JsonObject> transactions, String tenantId) {
     return String.format("UPDATE %s AS transactions " +
       "SET jsonb = t.jsonb FROM (VALUES  %s) AS t (id, jsonb) " +
